modif mock-pokemon and pokemon-service before deploy

+++ src/services/pokemon-service.ts
@@ -1,18 +1,33 @@
 import Pokemon from "../models/pokemon";
+import POKEMONS from "../models/mock-pokemon";
  
 export default class PokemonService {
- 
+  
+  static pokemons:Pokemon[] = POKEMONS;
+
+  static isDev = (!process.env.NODE_ENV || process.env.NODE_ENV === 'development');
+
   static getPokemons(): Promise<Pokemon[]> {
-    return fetch('http://localhost:3001/pokemons')
-      .then(response => response.json())
-      .catch(error => this.handleError(error));
+    if (this.isDev) {
+      return fetch('http://localhost:3001/pokemons')
+        .then(response => response.json())
+        .catch(error => this.handleError(error));
+    }
+    return new Promise( resolve => {
+      resolve(this.pokemons);
+      });
   }
  
   static getPokemon(id: number): Promise<Pokemon|null> {
-    return fetch(`http://localhost:3001/pokemons/${id}`)
+    if (this.isDev) {
+      return fetch(`http://localhost:3001/pokemons/${id}`)
       .then(response => response.json())
       .then(data => this.isEmpty(data) ? null : data)
       .catch(error => this.handleError(error));
+    }
+    return new Promise(resolve => {
+      resolve(this.pokemons.find(pokemon => id === pokemon.id));
+    });
   }
  
   static isEmpty(data: Object): boolean {
@@ -20,46 +35,72 @@ export default class PokemonService {
   }
 
   static updatePokemon(pokemon: Pokemon): Promise<Pokemon> {
-    return fetch(`http://localhost:3001/pokemons/${pokemon.id}`, {
-    method: 'PUT',
-    body: JSON.stringify(pokemon),
-    headers: { 'Content-Type': 'application/json'}
-    })
-    .then(response => response.json() )
-    .catch(error => this.handleError(error));
-
+    if (this.isDev){
+      return fetch(`http://localhost:3001/pokemons/${pokemon.id}`, {
+        method: 'PUT',
+        body: JSON.stringify(pokemon),
+        headers: { 'Content-Type': 'application/json'}
+      })
+      .then(response => response.json() )
+      .catch(error => this.handleError(error));
+    }
+    return new Promise(resolve => {
+      const { id } = pokemon;
+      const index = this.pokemons.findIndex(pokemon => pokemon.id === id)
+      this.pokemons[index] = pokemon;
+      resolve(pokemon);
+    });
   }
+
   static deletePokemon(pokemon: Pokemon): Promise<{}> {
-    return fetch(`http://localhost:3001/pokemons/${pokemon.id}`, {
-    method: 'DELETE',
-    body: JSON.stringify(pokemon),
-    headers: { 'Content-Type': 'application/json'}  
-    })
-    .then(response => response.json())
-    .catch(error => this.handleError(error));
+    if (this.isDev){
+      return fetch(`http://localhost:3001/pokemons/${pokemon.id}`, {
+        method: 'DELETE',
+        headers: { 'Content-Type': 'application/json'}  
+      })
+      .then(response => response.json())
+      .catch(error => this.handleError(error));
+    }
+    return new Promise(resolve => {
+      const { id } = pokemon;
+      this.pokemons = this.pokemons.filter(pokemon => pokemon.id !== id);
+      resolve({});
+    });
   }
 
   static addPokemon(pokemon:Pokemon): Promise<Pokemon> {
-    delete pokemon.created;
+    pokemon.created = new Date(pokemon.created);
     
-    const requestOptions = {
-      method: 'POST',
-      body: JSON.stringify(pokemon),
-      headers: { 'Content-Type': 'application/JSON'}
-    };
-    return fetch('http://localhost:3001/pokemons', requestOptions)
-    .then(response => response.json())
-    .catch(error => this.handleError(error));
+    if (this.isDev){
+      const requestOptions = {
+        method: 'POST',
+        body: JSON.stringify(pokemon),
+        headers: { 'Content-Type': 'application/JSON'}
+      };
+      return fetch('http://localhost:3001/pokemons', requestOptions)
+      .then(response => response.json())
+      .catch(error => this.handleError(error));
+    }
+    return new Promise(resolve => {
+      this.pokemons.push(pokemon);
+      resolve(pokemon);
+    });
   }
 
   static searchPokemon(term:string): Promise<Pokemon[]> {
-    return fetch(`http://localhost:3001/pokemons?q=${term}`)
+    if (this.isDev){
+      return fetch(`http://localhost:3001/pokemons?q=${term}`)
       .then(response => response.json())
       .catch(error => this.handleError(error));
+    }
+    return new Promise(resolve => {
+      const results = this.pokemons.filter((pokemon => pokemon.name.includes(term)) || (pokemon => pokemon.types.includes(term)));
+      resolve(results);
+    });
   }
+
 
   static handleError(error:Error): void {
       console.error(error);
   }
-  
 }
